Fix Directory.EnumerableTests
